Refactor database schema and models: move `Namespace` class and related models to `models.py`, update imports, and enhance structure for better organization.
