+++ src/database/models.py
@@ -1,56 +1,9 @@
-import abc
 import logging
-from typing import Any
 
 from pydantic import BaseModel
 
 from src.config import settings
-from src.database.db_connector import ConnectionCM, ConnectionType
-
-
-def create_all(cm_manager: ConnectionCM) -> None:
-    """ Creating required tables and objects in the assigned database and default data """
-    with cm_manager as connection:
-        for db_cls in DatabaseObject.__subclasses__():
-            db_instance = db_cls(connection)
-            db_instance.execute_ddl()
-            logging.info(f"{db_instance.name}: DDL executed")
-            if default_data := db_instance.default_data():
-                connection.execute(default_data)
-                connection.commit()
-                logging.info(f"{db_instance.name}: default data inserted")
-
-
-class DatabaseObject[PartM: BaseModel, FullM: BaseModel](abc.ABC):
-    """ Defines interface for required database objects (because SQLAlchemy & alembic support DuckDB badly) """
-    default_schema = settings.database.default_schema
-
-    name: str
-    autoincrement: str
-
-    def __init__(self, connection: ConnectionType):
-        self.connection = connection
-
-    def insert(self, model: PartM) -> FullM:
-        raise NotImplementedError
-
-    def get(self, id_: Any) -> FullM | None:
-        raise NotImplementedError
-
-    def update(self, model: FullM) -> FullM:
-        raise NotImplementedError
-
-    def delete(self, id_: Any) -> None:
-        raise NotImplementedError
-
-    def all(self) -> list[FullM]:
-        raise NotImplementedError
-
-    def execute_ddl(self) -> None:
-        raise NotImplementedError
-
-    def default_data(self) -> str | None:
-        ...
+from src.database.base_model import DatabaseObject
 
 
 class NamespacePartModel(BaseModel):
